requirements.txt
alembic==1.13.1
annotated-types==0.7.0
anyio==4.4.0
APScheduler==3.10.4
async-timeout==4.0.3
asyncpg==0.29.0
bcrypt==4.0.1
black==24.8.0
certifi==2024.6.2
click==8.1.7
dependency-injector==4.41.0
Deprecated==1.2.14
dnspython==2.6.1
ecdsa==0.19.0
email_validator==2.1.2
exceptiongroup==1.2.1
fastapi==0.111.0
fastapi-cli==0.0.4
flake8==7.1.1
greenlet==3.0.3
h11==0.14.0
httpcore==1.0.5
httptools==0.6.1
httpx==0.27.0
idna==3.7
importlib_metadata==8.0.0
Jinja2==3.1.4
jose==1.0.0
Mako==1.3.5
markdown-it-py==3.0.0
MarkupSafe==2.1.5
mccabe==0.7.0
mdurl==0.1.2
mypy-extensions==1.0.0
opentelemetry-api==1.26.0
opentelemetry-instrumentation==0.47b0
opentelemetry-instrumentation-sqlalchemy==0.47b0
opentelemetry-semantic-conventions==0.47b0
orjson==3.10.5
packaging==24.1
passlib==1.7.4
pathspec==0.12.1
platformdirs==4.2.2
psycopg2-binary==2.9.9
pyasn1==0.6.0
pycodestyle==2.12.1
pydantic==2.8.2
pydantic-settings==2.4.0
pydantic_core==2.20.1
pyflakes==3.2.0
Pygments==2.18.0
PyJWT==2.8.0
python-dotenv==1.0.1
python-jose==3.3.0
python-multipart==0.0.9
pytz==2024.1
PyYAML==6.0.1
rich==13.7.1
rsa==4.9
shellingham==1.5.4
six==1.16.0
sniffio==1.3.1
SQLAlchemy==2.0.31
starlette==0.37.2
tomli==2.0.1
typer==0.12.3
typing_extensions==4.12.2
tzlocal==5.2
ujson==5.10.0
uvicorn==0.30.5
uvloop==0.19.0
watchfiles==0.22.0
websockets==12.0
wrapt==1.16.0
zipp==3.19.2
